requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
httpx==0.27.2

pydantic==2.9.2

langchain==0.3.7
langchain-community==0.3.7
langchain-gigachat==0.3.12
langchain-text-splitters==0.3.2

chromadb==0.5.18
